Add #current_position
∆ Add unit test, state and behaviour to have a current position. In
order to have a position that can change, it needs a current position
to begin with. Currently the MarsRover class also includes the
behaviour to return this position as a string since this is required
later. Not comfortable with this class managing a presentation string,
but until a presentation class is constructed, this will have to
remain here.