Add #check_rover_position
∆ Add unit test and behaviour to check the position of a landed rover.
The next failure of US6 requires a check of a rovers position and a
return of "N", first step to this is to check the position, which can
now be done, but returns an empty string. This is because #land is
putting the rover on the plateau but not setting it's heading.