Add #change_rover_position
∆ Add unit test and behaviour to controller for a change of position.
The rotate_right method required in the feature test will be from the
MarsRover which I feel should be managed by the controller. The
other methods before returning the rovers current position. The first
delegation is to split the input into an array, the second is to loop
through the array and perform a MarsRover.right_turn each time an "R"
is seen in the input.